requirements.txt
sqlalchemy>=2.0
pyserial
psycopg2-binary
celery
redis
click
uvicorn[standard]
fastapi